pyproject.toml
[project]
name = "korus-personal-info-chekcer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "holidays>=0.73",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "xlrd>=2.0.1",
]
